refactor(campaign): decompose contractMarket.ts into 4 modules
Splits the 682-LOC contractMarket.ts (over the 600-LOC critical
threshold) into cohesive co-located sibling modules under
src/lib/campaign/contracts/. Behavior-preserving: all public exports
keep identical names and signatures, re-exported from contractMarket.ts
so every importer and test is unaffected.

New modules:
- contractMarketConstants.ts - frozen lookup tables + tuning constants
- contractRandomHelpers.ts   - RandomFn type + seeded random primitives
- contractSelection.ts       - per-attribute random selectors, name/id
                               generation, followup type selection
- contractPayment.ts         - shared payment-terms + contract-record
                               assembly helpers (removes 4x/3x dup)

contractMarket.ts drops from 682 to 391 LOC and now holds only the
orchestration functions (calculateForceBV, generateContracts,
generateAtBContracts, acceptContract, generateFollowupContract,
generateContractsWithStanding).

Verification: tsc --noEmit clean, oxlint 0 warnings/0 errors,
2106/2106 campaign jest tests pass.
